+++ src/rx-document.js
@@ -229,6 +229,7 @@ export class RxDocument {
         if (valueObj === null) return;
 
         let pathProperties = this.collection.schema.getSchemaByObjectPath(objPath);
+        if(typeof pathProperties === 'undefined') return;
         if (pathProperties.properties) pathProperties = pathProperties.properties;
 
         Object.keys(pathProperties)